+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+ - **Color Animator Parameter**: Manipulates quadruples of float parameter in Animators as Colors.
+
 ## [0.2.0] - 2023-02-12
 
 ### Added